Prepare replication by fixing tests
