+++ experiments/test/parsing/data_parallel_parsed.yaml
@@ -11,5 +11,5 @@ dataset: fashion-mnist
 max_epoch: 44
 learning_rate: 1e3
 learning_rate_decay: 0.0002
-seed: 142
+seed: 2053695854357871005
 loss: MSELoss
