+++ tests/util/config/test_learning_config.py
@@ -1,3 +1,4 @@
+import random
 import unittest
 from dataclasses import fields
 from pathlib import Path
@@ -42,14 +43,14 @@ class FedLearningConfigTest(unittest.TestCase):
                 self.assertNotEqual(getattr(self.default, field.name), getattr(self.learning_params, field.name), msg=field.name)
 
     def test_parsed_equals(self):
-        # TODO: Make possible to parse without experiment parser.
-        description = ExperimentParser(config_path=Path(TEST_FED_CONF)).parse()[0]
+        description = ExperimentParser(config_path=Path(TEST_FED_CONF)).parse()
+        job_description = description.train_tasks[0]
         train_task = TrainTask(identity='test_fed',
-                  job_parameters=description.job_class_parameters,
-                  priority=description.priority,
-                  experiment_config=description.get_experiment_configuration(),
-                  replication=-1,
-                  experiment_type=description.experiment_type)
+                               job_parameters=job_description.job_class_parameters[0],
+                               priority=None,
+                               replication=-1,
+                               experiment_type=job_description.experiment_type,
+                               seed=431)
         arrival_task = FederatedArrivalTask.build(Arrival(None, train_task, 'test_fed'), train_task.identifier, -1)
         template = render_template(arrival_task, 'Master', -1, TEST_FED_CONF)
 
@@ -63,19 +64,19 @@ class DistLearningConfigTest(unittest.TestCase):
 
 
     def setUp(self):
+        random.seed(42)
         self.learning_params = FedLearningConfig.from_yaml(Path(TEST_PARAM_CONF_DISTRIBUT))
 
 
     def test_parsed_equals(self):
-        description = ExperimentParser(config_path=Path(TEST_DIST_CONF)).parse()[0]
+        description = ExperimentParser(config_path=Path(TEST_DIST_CONF)).parse().train_tasks[0]
         train_task = TrainTask(identity='test_fed',
-                  job_parameters=description.job_class_parameters,
-                  priority=description.priority,
-                  experiment_config=description.get_experiment_configuration(),
+                  job_parameters=description.job_class_parameters[0],
+                  priority=None,
                   replication=-1,
-                  experiment_type=description.experiment_type)
+                  experiment_type=description.experiment_type,
+                  seed=2053695854357871005)
         arrival_task = DistributedArrivalTask.build(Arrival(None, train_task, 'test_fed'), train_task.identifier, -1)
         template = render_template(arrival_task, 'Master', -1, TEST_FED_CONF)
-
         self.assertEquals(DistLearningConfig.from_yaml(Path(TEST_PARSED_CONF_DIST)),
                           DistLearningConfig.from_dict(yaml.load(template, Loader=get_safe_loader())))
